[STF][BUILD] Adjust project setup and add StartupSarosSTF
The new StartupSarosSTF class will start the STF from the stf project. Due to the refactoring in the commits before I had to adjust some config files.
