+++ de.fu_berlin.inf.dpp.stf/META-INF/MANIFEST.MF
@@ -1,11 +1,14 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Saros STF
-Bundle-SymbolicName: de.fu_berlin.inf.dpp.stf
+Bundle-SymbolicName: de.fu_berlin.inf.dpp.stf;singleton:=true
 Bundle-Version: 0.1.0.DEVEL
 Bundle-Vendor: FU-Berlin
+Bundle-Activator: de.fu_berlin.inf.dpp.Saros
+Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Require-Bundle: de.fu_berlin.inf.dpp.core,
+Require-Bundle: de.fu_berlin.inf.dpp,
+ de.fu_berlin.inf.dpp.core,
  de.fu_berlin.inf.dpp.ui,
  org.eclipse.swt
 Bundle-ClassPath: .
