+++ de.fu_berlin.inf.dpp.stf/plugin.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="4.6"?>
+ 
+  <plugin>
+    <extension
+          id="de.fu_berlin.inf.dpp.stf"
+          point="org.eclipse.ui.startup">
+       <startup
+             class="de.fu_berlin.inf.dpp.stf.StartupSarosSTF">
+       </startup>
+    </extension>
+ </plugin>
