+++ de.fu_berlin.inf.dpp/test/resources/launch/Saros_STF_Carl.launch
@@ -16,7 +16,7 @@
 <stringAttribute key="org.eclipse.jdt.launching.JRE_CONTAINER" value="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 <stringAttribute key="org.eclipse.jdt.launching.PROGRAM_ARGUMENTS" value="-os ${target.os} -ws ${target.ws} -arch ${target.arch} -nl en_US"/>
 <stringAttribute key="org.eclipse.jdt.launching.SOURCE_PATH_PROVIDER" value="org.eclipse.pde.ui.workbenchClasspathProvider"/>
-<stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value="-Xmx512m -ea -XX:MaxPermSize=256m -Djava.security.manager &quot;-Djava.security.policy=file:${workspace_loc:de.fu_berlin.inf.dpp}/test/resources/stf/all.policy&quot;  -Dde.fu_berlin.inf.dpp.debug=true -Dde.fu_berlin.inf.dpp.testmode=12347 -Dde.fu_berlin.inf.dpp.sleepTime=200 -Djava.rmi.server.hostname=localhost -Dorg.eclipse.swtbot.keyboard.strategy=org.eclipse.swtbot.swt.finder.keyboard.MockKeyboardStrategy  -Dorg.eclipse.swtbot.keyboard.layout=de.fu_berlin.inf.dpp.stf.server.bot.default -Duser.language=en -Duser.country=US -Dfile.encoding=UTF-8 -Djava.util.Arrays.useLegacyMergeSort=true"/>
+<stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value="-Xmx512m -ea -XX:MaxPermSize=256m -Djava.security.manager &quot;-Djava.security.policy=file:${workspace_loc:de.fu_berlin.inf.dpp.stf}/test/resources/all.policy&quot;  -Dde.fu_berlin.inf.dpp.debug=true -Dde.fu_berlin.inf.dpp.testmode=12347 -Dde.fu_berlin.inf.dpp.sleepTime=200 -Djava.rmi.server.hostname=localhost -Dorg.eclipse.swtbot.keyboard.strategy=org.eclipse.swtbot.swt.finder.keyboard.MockKeyboardStrategy  -Dorg.eclipse.swtbot.keyboard.layout=de.fu_berlin.inf.dpp.stf.server.bot.default -Duser.language=en -Duser.country=US -Dfile.encoding=UTF-8 -Djava.util.Arrays.useLegacyMergeSort=true"/>
 <stringAttribute key="org.eclipse.jdt.launching.WORKING_DIRECTORY" value="${workspace_loc}/../working-dir-carl-stf"/>
 <stringAttribute key="pde.version" value="3.3"/>
 <stringAttribute key="product" value="org.eclipse.platform.ide"/>
